requirements.txt
altair
numpy
pandas
pydeck
streamlit
matplotlib
plotly
scikit-learn
statsmodels
seaborn
umap-learn
hdbscan
shap
optuna
imblearn